Funcion de añadir semanas creada + copiar la semana actual de 3 formas diferentes.

+++ app/src/main/java/com/cursointermedio/myapplication/domain/model/WeekModel.kt
@@ -4,8 +4,8 @@ import com.cursointermedio.myapplication.data.database.entities.WeekEntity
 
 data class WeekModel
     (
-    val weekId: Int?,
-    val trainingWeekId: Int,
+    val weekId: Long?,
+    val trainingWeekId: Long,
     val name: String?,
     val description: String?
 )
